move xmlerror into "exceptions" module
