+++ src/dsp_tools/models/exceptions.py
@@ -75,3 +75,15 @@ class UserError(BaseError):
     due to a bad condition in the input data that prevents further processing.
     The message should be as user-friendly as possible.
     """
+
+
+class XmlError(Exception):
+    """Represents an error raised in the context of the XML import"""
+
+    _message: str
+
+    def __init__(self, msg: str):
+        self._message = msg
+
+    def __str__(self) -> str:
+        return "XML-ERROR: " + self._message
